wip: move Watcher et al to cdp/event package
