+++ cdp/event/event.go
@@ -1,0 +1,6 @@
+package event
+
+type Event struct {
+	Name CDPName
+	Data interface{}
+}
