+++ cdp/event/page.go
@@ -1,0 +1,7 @@
+package event
+
+type CDPName string
+
+const (
+	PageNavigated CDPName = "Page.navigated"
+)
